refactor: polishing

+++ src/main/resources/data.sql
@@ -4,26 +4,26 @@ DROP TABLE GAME IF EXISTS;
 
 CREATE TABLE GAME
 (
-    id          BIGINT         NOT NULL AUTO_INCREMENT,
-    created_at  DATETIME    NOT NULL default current_timestamp,
-    trial_count INT         NOT NULL,
+    id          BIGINT   NOT NULL AUTO_INCREMENT,
+    created_at  DATETIME NOT NULL default current_timestamp,
+    trial_count INT      NOT NULL,
     PRIMARY KEY (id)
 );
 
 
 CREATE TABLE CAR
 (
-    id       BIGINT         NOT NULL AUTO_INCREMENT,
+    id       BIGINT      NOT NULL AUTO_INCREMENT,
     name     VARCHAR(50) NOT NULL,
     position INT         NOT NULL,
-    game_id  INT         NOT NULL,
+    game_id  BIGINT         NOT NULL,
     PRIMARY KEY (id)
 );
 
 CREATE TABLE WINNER
 (
-    id   BIGINT         NOT NULL AUTO_INCREMENT,
-    name VARCHAR(50) NOT NULL,
-    game_id  INT         NOT NULL,
+    id      BIGINT      NOT NULL AUTO_INCREMENT,
+    name    VARCHAR(50) NOT NULL,
+    game_id BIGINT         NOT NULL,
     PRIMARY KEY (id)
 )
